close offline player loophole
It seems like this code is a leftover from when jds didn't
have digilines yet and the player was the one interacting
with the formspec. Now the player is the owner of the jd.

Also back then can_travel() was inspecting XP instead of
privs. XP is stored in player meta, so isn't checkable
when player is offline.

With this change, players can't jump over warzone or other
places they normally couldn't go to (except landing-priv area). They
used an alt to place the JD then while the alt wasn't online, the checks
were skipped and by trapping the
online player in a chair(/bike/minecart/...) the
teleport_back globalstep couldn't force them
back to /spawn.